very big refactor, add elective_course router, add pydantic schemas, very big rewrite database, fix changes (probably it works, but now i cant be sure in this fact ): ),
