+++ src/tgbot/admin.py
@@ -3,10 +3,10 @@ from aiogram.filters import Command
 from aiogram.types import Message
 
 from config import ADMINS
-from models.database import get_async_session
-from models.db import DB
-from tgbot.handlers.auxiliary import bot
-from tgbot.text import TEXT
+from src.database import get_async_session
+from src.tgbot.auxiliary import bot
+from src.tgbot.text import TEXT
+from src.tgbot.user_models.db import DB
 
 router = Router()
