+++ src/tgbot/changes/schemas.py
@@ -1,0 +1,11 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class ChangesType(BaseModel):
+    id: Optional[int] = None
+    type: str
+    second: str
+    weekday: str
+    schedule: dict
