+++ src/tgbot/elective_course/schemas.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class ElectiveCourse(BaseModel):
+    name: str
+    pulpit: str
+    subject: str
+    timetable: dict
+    teacher_name: str
