+++ src/tgbot/user_models/db.py
@@ -1,33 +1,25 @@
-import json
-
 from sqlalchemy import select, insert, delete, update
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from models.model import users, changes, columns_json
-
-
-# продублировано из my_types
-class ChangesType:
-    def __init__(self, _type: str, _second: str, _weekday: str, _schedule: dict):
-        self.type = _type
-        self.second = _second
-        self.weekday = _weekday
-        self.schedule = _schedule
+from src.tgbot.user_models.models import UsersModel, columns_json
+from src.tgbot.user_models.schemas import User
 
 
 class DB:
     # Возвращает None если запись не найдется, иначе вернется dict
+    # TODO: сделать чтобы возвращалась схема данных
     async def select_user_by_id(self, session: AsyncSession, _id: int) -> dict | None:
         _id = self.__convert_to_id_type(_id)
-
-        query = select(users).where(users.c.id == _id)
-        res = await session.execute(query)
-        final_result = {}
+        query = select(UsersModel).where(UsersModel.id == _id)
 
         try:
-            for index, elem in enumerate(res.all()[0]):
-                if index == 0:
-                    elem = self.__convert_from_id_type(elem)
+            temp = await session.execute(query)
+            row_res = tuple(*temp)[0]
+
+            final_result = {}
+
+            for index, elem in enumerate([self.__convert_to_id_type(row_res.id), row_res.role, row_res.sub_info,
+                                          row_res.lang]):
                 final_result[columns_json[index]] = elem
         except IndexError:
             await session.commit()
@@ -36,24 +28,26 @@ class DB:
         await session.commit()
         return final_result
 
-    async def select_users_by_role_and_sub_info(self, session: AsyncSession, role: str, sub_info: str) -> list[dict]:
-        query = select(users).where(users.c.role == role, users.c.sub_info == sub_info)
+    @staticmethod
+    async def select_users_by_role_and_sub_info(session: AsyncSession, role: str, sub_info: str) -> list[User]:
+        query = select(UsersModel).where(UsersModel.role == role, UsersModel.sub_info == sub_info)
 
         res = await session.execute(query)
         final_result = []
-
-        for i, user in enumerate(res.all()):
-            final_result.append({})
-            for index, elem in enumerate(user):
-                if index == 0:
-                    elem = self.__convert_from_id_type(elem)
-                final_result[i][columns_json[index]] = elem
+        for user in res.all():
+            final_result.append(User(**user[0].__dict__))
+            # final_result.append(**user)
+            # final_result.append({})
+            # for index, elem in enumerate(user):
+            #     if index == 0:
+            #         elem = self.__convert_from_id_type(elem)
+            #     final_result[i][columns_json[index]] = elem
 
         await session.commit()
         return final_result
 
     async def get_all_users(self, session: AsyncSession) -> list[dict]:
-        query = select(users)
+        query = select(UsersModel)
 
         res = await session.execute(query)
         final_result = []
@@ -69,7 +63,7 @@ class DB:
         return final_result
 
     async def create_user(self, session: AsyncSession, **kwargs):
-        # what must be in kwargs u can see in models.py
+        # what must be in kwargs u can see in user_models.py
         # проверка, что переданы все параметры
         if list(kwargs.keys()) != list(columns_json.values()):
             raise ValueError('Не хватает параметров для создания пользователя')
@@ -77,21 +71,21 @@ class DB:
         # преобразование id в тип id, который находтся в бд
         kwargs['id'] = self.__convert_to_id_type(kwargs['id'])
 
-        stmt = insert(users).values(**kwargs)
+        stmt = insert(UsersModel).values(**kwargs)
         await session.execute(stmt)
         await session.commit()
 
     async def delete_user(self, session: AsyncSession, _id):
         _id = self.__convert_to_id_type(_id)
 
-        stmt = delete(users).where(users.c.id == _id)
+        stmt = delete(UsersModel).where(UsersModel.id == _id)
         await session.execute(stmt)
         await session.commit()
 
     async def update_user_info(self, session: AsyncSession, _id, **kwargs):
         _id = self.__convert_to_id_type(_id)
 
-        stmt = update(users).where(users.c.id == _id).values(**kwargs)
+        stmt = update(UsersModel).where(UsersModel.id == _id).values(**kwargs)
 
         await session.execute(stmt)
         await session.commit()
@@ -104,7 +98,6 @@ class DB:
     def __convert_from_id_type(_id) -> int:
         return int(_id)
 
-
 # SAMPLE USAGE
 # async def main():
 #     session = await get_async_session()
@@ -113,33 +106,3 @@ class DB:
 #
 # # Run the main function
 # asyncio.run(main())
-
-
-class ChangesDB:
-    @staticmethod
-    async def add_changes(session: AsyncSession, changed_schedule: ChangesType):
-        stmt = insert(changes).values((changed_schedule.type, changed_schedule.second, changed_schedule.weekday,
-                                       json.dumps(changed_schedule.schedule)))
-
-        await session.execute(stmt)
-        await session.commit()
-
-    @staticmethod
-    async def get_all_changes(session: AsyncSession) -> list[ChangesType]:
-        query = select(changes.c.type, changes.c.second, changes.c.weekday, changes.c.schedule)
-
-        query_result = await session.execute(query)
-        await session.commit()
-        result = []
-
-        for i in query_result.all():
-            result.append(ChangesType(i[0], i[1], i[2], json.loads(i[3])))
-
-        return result
-
-    @staticmethod
-    async def delete_all_changes(session: AsyncSession):
-        stmt = delete(changes)
-
-        await session.execute(stmt)
-        await session.commit()
