+++ src/tgbot/user_models/schemas.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+# TODO: сделать более продвинутую модель
+class User(BaseModel):
+    id: int
+    role: str
+    sub_info: str
+    lang: str
